refactor(prospect): stage-decompose createProspect, no behavior change (P3-1)
createProspect had grown to ~1,200 lines by inlining every stage of lead
capture. This lifts out the two stages with the cleanest seams, verbatim:

  PERSIST  → prospectCreateTx.js   the single create transaction
  DISPATCH → prospectDispatch.js   everything post-commit

Both were closures reaching ~20 enclosing variables each. Every input is now a
named ctx field and every value the caller needs back is a named result field,
so the data flow across each boundary is readable without holding the whole
function in your head. prospectService.js drops 1,736 → 1,307 lines and
createProspect drops ~1,200 → ~780.

Nothing about the behaviour changes. The statement order inside each stage is
untouched, including the parts that are load-bearing: the quota/DNC/screening
gate precedence, the savepoint isolation around the consumer spine and the
enrichment outbox, and the post-commit ordering that keeps DNC scrubbing ahead
of the webhook and the screening dial last. The only delta is that
dncIntendedAgentId and dncAlreadyCharged are now locals — the caller never read
them.

Proof is the existing suites passing with ZERO test edits: unit 129/129 suites
(2,202 tests) and the full integration tier 147/147 suites (2,593 tests).

The new unit test is coverage that could not exist before: as closures these
stages were reachable only by driving a full HTTP capture against Postgres, so
gate precedence and dispatch suppression had no direct test. 13 cases, no
database.
