+++ backend/src/services/prospectService.js
@@ -17,13 +17,15 @@ import { makeIdempotencyOps } from './idempotencyProtocol.js';
 import { makeProspectReads } from './prospectReadService.js';
 import { makeHeldQueueOps } from './prospectHeldQueueService.js';
 import { makeAssignmentOps } from './prospectAssignmentService.js';
+import { makeCreateTxRunner } from './prospectCreateTx.js';
+import { makeDispatchRunner } from './prospectDispatch.js';
 import {
   UUID_RE, PROSPECT_UPDATE_FIELDS,
 } from './prospectShared.js';
 import { deductLeadCredit, chargeLeadCredit, deductExternalLeadBalance } from './leadCredits.js';
 import { decideAssignment } from './leadQuota.js';
 import { dncEnforcement, formatDncNumber, checkAndRecord as dncCheckAndRecord } from './dncService.js';
-import { gateHeldDncLead, dncCaptureGate, bakeHoldTargetAgentId } from './dncGate.js';
+import { gateHeldDncLead, dncCaptureGate } from './dncGate.js';
 import { hasValidExternalConsent, buildExternalConsentEvidence } from './externalConsent.js';
 import { buildDncConsentEvidence } from './dncConsent.js';
 import { buildProspectWhere } from './prospectScope.js';
@@ -74,14 +76,10 @@ export function registerLeadCapturedHook(fn) {
   _leadCapturedHook = typeof fn === 'function' ? fn : null;
 }
 import { logger } from '../utils/logger.js';
-import { signupActivityDescription } from '../utils/sourceLabel.js';
 import {
   normalizePhone,
-  buildLeadCreatedPayload,
-  buildLeadHeldPayload,
   buildLeadDeletedPayload,
   destinationForAgent,
-  externalIdForDestination,
 } from './prospectHelpers.js';
 import { scoreQuiz } from './quizScoringService.js';
 import { readLegacyViewSafe } from '../utils/designConfigV2Clamp.js';
@@ -156,6 +154,10 @@ export function makeProspectService(overrides = {}) {
   const { assignProspect, bulkAssignProspects } = assigns;
   const held = makeHeldQueueOps({ d, m, idem, assignProspect });
   const { orphanSystemAgentId, listDispatchableOrphans, releaseHeldProspect, reassignProspectExternal, returnProspectToHeld, bulkReturnProspectsToHeld } = held;
+  // The persist stage of createProspect (P3-1) — see prospectCreateTx.js.
+  const runCreateTx = makeCreateTxRunner({ d, m });
+  // The post-commit stage of createProspect (P3-1) — see prospectDispatch.js.
+  const runDispatch = makeDispatchRunner({ d, m });
 
   /**
    * Create a new prospect (lead capture).
@@ -844,252 +846,34 @@ export function makeProspectService(overrides = {}) {
     //     lead, and for a funded gated route charges a credit authoritatively
     //     (charged:true ⇒ skip the best-effort deduct below to avoid double-charging).
     //     Soft/exempt routes are unchanged: assign + best-effort deduct.
-    let quarantined = false;
-    let heldReason = null;
-    let finalAgentId = assignedAgentId;
-    // DNC block-mode hold bookkeeping — set inside the tx, consumed post-commit.
-    let dncHeld = false;
-    let dncIntendedAgentId = null;
-    let dncAlreadyCharged = false;
-    // Screening hold bookkeeping (plan §5.2) — mirrors the DNC shape.
-    let screeningHeld = false;
-    const runCreateTx = () => d.sequelize.transaction(async (t) => {
-      // The internal quota gate applies ONLY to the internal path. For external
-      // leads default to a plain "assign" directive so the shared activity/deduct
-      // code below stays correct (external is charged authoritatively, not metered).
-      let decision = { action: 'assign', assignedAgentId, charged: false, via: routeVia };
-      if (externalHold) {
-        // External-eligible + consented but no funded buyer → HOLD (never System Agent,
-        // never charged). The distinct quarantineReason fences this lead off from the
-        // internal release sweep so it can never be delivered to Lyfe.
-        decision = { action: 'quarantine', quarantineReason: externalHoldReason, charged: false, via: routeVia };
-      } else if (!externalAgentId) {
-        decision = await d.decideAssignment({
-          campaign: sourceCampaign,
-          routing: { agentId: assignedAgentId, via: routeVia },
-          campaignId: incoming.campaignId,
-          transaction: t,
-          charge: d.chargeLeadCredit,
-        });
-      }
-
-      // Hold-target bake rule — shared with the Retell capture path (P1-3);
-      // the rule itself is documented on bakeHoldTargetAgentId in dncGate.js.
-      const bakeIntendedAgentId = (candidateId) =>
-        bakeHoldTargetAgentId(candidateId, { routeVia, User: m.User, transaction: t });
-
-      // DNC block-mode gate: a normally-assignable INTERNAL lead is HELD pending a DNC
-      // check (released post-commit on clear). Never overrides an existing quarantine
-      // (quota / external) or an external-buyer route. The credit is charged on release
-      // (unless decideAssignment already charged a funded gated route → dncAlreadyCharged).
-      if (dncBlockApplies && decision.action !== 'quarantine' && !externalAgentId) {
-        dncIntendedAgentId = await bakeIntendedAgentId(decision.assignedAgentId ?? assignedAgentId ?? null);
-        dncAlreadyCharged = decision.charged === true;
-        dncHeld = true;
-        decision = { action: 'quarantine', quarantineReason: 'dnc_pending', charged: dncAlreadyCharged, via: routeVia };
-      }
-
-      // Screening gate (plan §5.2): a normally-assignable INTERNAL lead is HELD
-      // pending the AI screening call. Sits AFTER the DNC block on purpose —
-      // when both gates apply the lead is born dnc_pending and dncGate hands
-      // off to screening on clear (never dial an unscrubbed number, §6).
-      let screeningIntendedAgentId = null;
-      let screeningAlreadyCharged = false;
-      if (screeningWanted && decision.action !== 'quarantine' && !externalAgentId) {
-        screeningIntendedAgentId = await bakeIntendedAgentId(decision.assignedAgentId ?? assignedAgentId ?? null);
-        screeningAlreadyCharged = decision.charged === true;
-        screeningHeld = true;
-        decision = { action: 'quarantine', quarantineReason: 'screening_pending', charged: screeningAlreadyCharged, via: routeVia };
-      }
-
-      quarantined = decision.action === 'quarantine';
-      heldReason = quarantined ? decision.quarantineReason : null;
-      finalAgentId = quarantined ? null : (decision.assignedAgentId ?? null);
-
-      // Consumer spine (docs/plans/consumer-spine-and-consent-ledger.md §2.3):
-      // resolve-or-create the cross-campaign person for this already-normalized
-      // phone INSIDE A SAVEPOINT — a spine failure rolls back the savepoint
-      // only and capture proceeds with consumerId null (reconciler heals).
-      // Runs AFTER the per-campaign 409 dupe check so duplicates never bump
-      // signupCount. call_bot never links (to_number ambiguity — see
-      // consumerService.js); `verified` is the single OTP-marker read above.
-      const consumerId = incoming.leadSource === 'call_bot'
-        ? null
-        : await d.resolveConsumerForCaptureTx(t, {
-            phone: incoming.phone,
-            firstName: incoming.firstName,
-            lastName: incoming.lastName,
-            email: incoming.email,
-            verified: otpMarkerLive,
-          });
-
-      const newProspect = await m.Prospect.create(
-        {
-          ...incoming,
-          consumerId,
-          assignedAgentId: finalAgentId,
-          externalAgentId,
-          quarantinedAt: quarantined ? new Date() : null,
-          quarantineReason: quarantined ? decision.quarantineReason : null,
-          // DNC: mark pending up-front (crash-safe — the backfill finds a stranded row);
-          // stash the intended agent so the post-commit clear-release knows who to deliver to.
-          ...(dncWillCheck ? { dncStatus: 'pending' } : {}),
-          ...(dncHeld ? { dncMetadata: { intendedAgentId: dncIntendedAgentId, alreadyCharged: dncAlreadyCharged } } : {}),
-          ...(screeningHeld
-            ? { screeningMetadata: { intendedAgentId: screeningIntendedAgentId, alreadyCharged: screeningAlreadyCharged, attempts: {} } }
-            : {}),
-        },
-        { transaction: t }
-      );
-
-      // Consent ledger (PR B, plan §3.1): person-level evidence for this
-      // capture — savepoint-isolated + non-blocking like the resolver; every
-      // input also lives on the prospect row, so backfillConsentEvents can
-      // re-derive a missed write. No-op when consumerId is null (call_bot /
-      // unlinked rows carry no person-level evidence until reconciled).
-      await d.recordCaptureConsentEventsTx(t, {
-        consumerId,
-        prospectId: newProspect.id,
-        campaignId: newProspect.campaignId || null,
-        sourceUrl: eventSourceUrl || null,
-        verified: otpMarkerLive,
-        contact: consentContact,
-        copyVersion: consentCopyVersion,
-        terms: consentTerms,
+    // --- PERSIST: the one transaction that turns this lead into rows ---
+    // The body lives in prospectCreateTx.js (P3-1). It used to be inlined here
+    // and reached ~20 enclosing variables through the closure; now every input
+    // is a named ctx field and every decision it makes is a named result field.
+    let txOutcome;
+    try {
+      txOutcome = await runCreateTx({
+        incoming,
+        user,
+        sourceCampaign,
+        sourceQrTag,
+        assignedAgentId,
+        externalAgentId,
+        externalHold,
+        externalHoldReason,
+        routeVia,
+        dncBlockApplies,
+        dncWillCheck,
+        screeningWanted,
+        otpMarkerLive,
+        eventSourceUrl,
+        consentContact,
+        consentTerms,
+        consentCopyVersion,
         externalConsent,
         dncConsent,
-        drawTerms: incoming.consentMetadata?.drawTerms || null,
+        acceptedProfileFacts,
       });
-
-      // Enrichment outbox (docs/plans/consumer-profile-enrichment.md §5):
-      // freeze the normalized fact snapshot and enqueue the map job IN THIS
-      // transaction (crash-safe — Codex R1 #5), savepoint-isolated like the
-      // spine resolver: enrichment must never fail capture; the nightly
-      // sweep re-drains anything a crash or savepoint rollback loses.
-      try {
-        await d.sequelize.transaction({ transaction: t }, async (sp) => {
-          const snapshot = d.buildFactSnapshot({
-            demographics: newProspect.demographics,
-            sourceMetadata: newProspect.sourceMetadata,
-            quizDefinition: sourceCampaign?.design_config?.quiz,
-            profileFacts: acceptedProfileFacts,
-          });
-          await d.enqueueMapJobsTx(sp, {
-            prospectId: newProspect.id,
-            formRevision: newProspect.enrichmentRevision || 1,
-            snapshot,
-          });
-        });
-      } catch (enrichErr) {
-        d.logger.warn('[enrichment] map-job outbox failed (sweep heals)', {
-          error: enrichErr?.message || String(enrichErr),
-        });
-      }
-
-      const campaignName = sourceCampaign?.name || 'Unknown Campaign';
-      // Source-aware phrase ("via TikTok ad" / "via web form" / "via {name} QR
-      // code" …) instead of the old hardcoded "via {qr} QR code", which
-      // mislabeled every non-QR lead as "Unknown QR". See utils/sourceLabel.js.
-      const activityDescription = signupActivityDescription(campaignName, {
-        leadSource: incoming.leadSource,
-        qrTag: sourceQrTag,
-        sourceMetadata: incoming.sourceMetadata,
-      });
-
-      // Activity: created
-      await m.ProspectActivity.create(
-        {
-          prospectId: newProspect.id,
-          type: 'created',
-          actorUserId: user?.id || null,
-          description: activityDescription,
-          metadata: {
-            leadSource: incoming.leadSource,
-            campaignId: newProspect.campaignId,
-            qrTagId: newProspect.qrTagId,
-          },
-        },
-        { transaction: t }
-      );
-
-      // Activity: assignment outcome (assigned, or held under quota)
-      if (quarantined) {
-        await m.ProspectActivity.create(
-          {
-            prospectId: newProspect.id,
-            type: 'updated',
-            actorUserId: user?.id || null,
-            description:
-              decision.quarantineReason === 'no_funded_external_buyer'
-                ? 'Held — no funded MKTR Leads (external) buyer'
-                : decision.quarantineReason === 'dnc_pending'
-                  ? 'Held — pending DNC (Do Not Call) check'
-                  : decision.quarantineReason === 'screening_pending'
-                    ? 'Held — pending AI screening call'
-                    : 'Held — no funded agent (lead quota)',
-            metadata: { quarantined: true, reason: decision.quarantineReason, via: routeVia },
-          },
-          { transaction: t }
-        );
-      } else {
-        await m.ProspectActivity.create(
-          {
-            prospectId: newProspect.id,
-            type: 'assigned',
-            actorUserId: user?.id || null,
-            description: externalAgentId
-              ? `Routed to external buyer ${externalAgentId} (MKTR Leads)`
-              : `Assigned to agent ${finalAgentId}`,
-            metadata: externalAgentId
-              ? { externalAgentId }
-              : { assignedAgentId: finalAgentId },
-          },
-          { transaction: t }
-        );
-
-        // Deduct lead credit.
-        //  - External (MKTR Leads) buyers are PAID leads: the charge is authoritative
-        //    — if the buyer's prepaid balance can't be charged, abort the whole create
-        //    (rollback) rather than hand over a lead we can't bill.
-        //  - Internal stays best-effort, and is skipped when decideAssignment already
-        //    charged authoritatively (charged:true) to avoid double-charging.
-        if (externalAgentId) {
-          const extCharged = await d.deductExternalLeadBalance(externalAgentId, 1, t);
-          if (!extCharged) {
-            throw new d.AppError('No paid external buyer balance available for this lead.', 409);
-          }
-        } else if (finalAgentId && decision.charged !== true) {
-          await d
-            .deductLeadCredit({ agentId: finalAgentId, campaignId: newProspect.campaignId || null, transaction: t })
-            .catch((err) => d.logger.error('Failed to deduct credit', { error: err?.message || String(err) }));
-        }
-      }
-
-      // Update QR tag analytics (atomic to avoid read-modify-write race).
-      // A quarantined lead is still a captured conversion, so we count it.
-      if (newProspect.qrTagId && sourceQrTag) {
-        await sourceQrTag.update(
-          {
-            analytics: d.sequelize.literal(`
-            jsonb_set(
-              COALESCE(analytics::jsonb, '{}'),
-              '{conversions}',
-              to_jsonb(COALESCE((analytics->>'conversions')::int, 0) + 1)
-            )
-          `),
-          },
-          { transaction: t }
-        );
-      }
-
-      // Campaign metrics are now computed from real data (no JSON blob to increment)
-
-      return newProspect;
-    });
-
-    let prospect;
-    try {
-      prospect = await runCreateTx();
     } catch (err) {
       // Concurrent same-campaign duplicate: both requests passed the precheck;
       // the unique-index loser lands here AFTER a full rollback (including the
@@ -1110,254 +894,41 @@ export function makeProspectService(overrides = {}) {
       throw err;
     }
 
+    const { prospect, quarantined, heldReason, finalAgentId, dncHeld, screeningHeld } = txOutcome;
+
     // Reflect the committed outcome (null when quarantined) for the rest of the
     // function — webhook dispatch, agent load, and the returned payload.
     assignedAgentId = finalAgentId;
 
-    // --- DNC scrubbing (post-commit, synchronous before dispatch) ---
-    // Block mode: the lead was born held (dnc_pending). Check, then release-on-clear (which
-    // fires its own first lead.created) or keep held (dnc_registered) — so the normal
-    // lead.created below stays suppressed (quarantined). Flag mode: check + record so the
-    // lead.created payload below carries the result; the lead delivers regardless.
-    if (dncHeld) {
-      await d.gateHeldDncLead(prospect).catch((err) =>
-        d.logger.error('[DNC] gate error', { error: err?.message || String(err) })
-      );
-    } else if (dncFlagApplies) {
-      await d
-        .dncCheckAndRecord(prospect)
-        .catch((err) => d.logger.error('[DNC] check error', { error: err?.message || String(err) }));
-    }
-
-    // --- Webhook dispatch (AFTER transaction commits, fire-and-forget) ---
-    // Always load the assigned agent's provenance (lyfeId/mktrLeadsId) by id — NOT
-    // the possibly-partial resolvedAgent from QR/group routing, which lacks it — so
-    // we route to the right app and send the destination-correct external id.
-    let agentForWebhook = null;
-    let leadDestination = null;
-    if (assignedAgentId) {
-      const agentRecord = await m.User.findByPk(assignedAgentId, {
-        attributes: ['id', 'lyfeId', 'mktrLeadsId', 'phone', 'email', 'firstName', 'lastName'],
-      });
-      if (agentRecord) {
-        leadDestination = destinationForAgent(agentRecord);
-        agentForWebhook = {
-          phone: agentRecord.phone || resolvedAgent?.phone || null,
-          email: agentRecord.email || resolvedAgent?.email || null,
-          name: `${agentRecord.firstName || ''} ${agentRecord.lastName || ''}`.trim() || resolvedAgent?.name || null,
-          id: externalIdForDestination(agentRecord, leadDestination),
-        };
-      }
-    }
-
-    // Suppress the Lyfe 'lead.created' delivery webhook when there is no internal
-    // agent to deliver to:
-    //  - quarantined (held under lead quota) — no agent yet; it fires on release
-    //    (slice 4) as the first lead.created.
-    //  - external (MKTR Leads) — the existing subscriber is the Lyfe app, and an
-    //    external buyer lead must NEVER be dispatched to it (no external subscriber
-    //    exists yet; destination-aware routing lands later in webhookService).
-    if (!quarantined && !externalAgentId) {
-      d.dispatchEvent('lead.created', () =>
-        buildLeadCreatedPayload(
-          prospect,
-          routingMode,
-          agentForWebhook,
-          assignedAgentId,
-          sourceCampaign,
-          sourceQrTag,
-          agentGroup
-        ),
-        { destination: leadDestination }
-      ).catch((err) => {
-        d.logger.error('[Webhook] dispatch error', { error: err?.message || String(err) });
-      });
-      // Suppressed-person new-lead propagation rides webhookService's
-      // flush-time catchup choke point (tracker "propagate") — no per-site
-      // hook needed here.
-    }
-
-    // Held (no_funded_agent) → ping the mktr-leads admin held queue so a pending
-    // lead is never silent. ONLY this reason: the external (no_funded_external_buyer)
-    // hold is a DIFFERENT, fenced pool that is NOT in that admin queue, so it must
-    // not ping. Gated by HELD_LEAD_PING_ENABLED; the sweep is the completeness net.
-    if (
-      quarantined &&
-      heldReason === 'no_funded_agent' &&
-      String(process.env.HELD_LEAD_PING_ENABLED || 'false').toLowerCase() === 'true'
-    ) {
-      d.dispatchEvent('lead.held', () => buildLeadHeldPayload(prospect, sourceCampaign, heldReason), {
-        destination: 'mktr_leads',
-      }).catch((err) => {
-        d.logger.error('[Webhook] lead.held dispatch error', { error: err?.message || String(err) });
-      });
-    }
-
-    // em/ph gate for every submit-time CAPI dispatch (Meta + TikTok), derived
-    // from the consent ledger ONCE — the capture txn just committed, so the
-    // capture-hook contact event (with its OTP `verified` stamp) is visible.
-    // FAIL CLOSED on any lookup error: the events still fire, without em/ph.
-    // The stored consent_contact boolean keeps being WRITTEN at capture
-    // (evidence/backfill) but is no longer read here (3sites).
-    let capiMarketingConsent = false;
-    try {
-      capiMarketingConsent = (await d.canMarketTo({
-        consumerId: prospect.consumerId || null,
-        phone: prospect.phone || null,
-        channel: 'all',
-        campaignId: prospect.campaignId || null,
-      })) === true;
-    } catch (err) {
-      d.logger.warn('[CAPI] canMarketTo failed — omitting em/ph (fail-closed)', {
-        error: err?.message || String(err),
-      });
-    }
-
-    // Meta CAPI dispatch (fire-and-forget; post-commit; guard inside sendLeadEvent)
-    d.sendLeadEvent(prospect, {
+    // --- DISPATCH: everything post-commit ---
+    // The body lives in prospectDispatch.js (P3-1): DNC scrubbing, the Lyfe
+    // webhook + held ping, the Meta/TikTok CAPI pairs, the Redeem Ops hook, the
+    // enrichment drain, the share link and the screening dial. All of it is
+    // fire-and-forget by rule, and its statement order is load-bearing.
+    const { assignedAgent, prospectWithCampaign, shareUrl, leadCapturedOutcome } = await runDispatch({
+      prospect,
+      quarantined,
+      heldReason,
+      assignedAgentId,
+      dncHeld,
+      dncFlagApplies,
+      screeningHeld,
+      externalAgentId,
+      sourceCampaign,
+      sourceQrTag,
+      resolvedAgent,
+      agentGroup,
+      routingMode,
       eventId,
+      registrationEventId,
+      eventSourceUrl,
       fbp,
       fbc,
-      eventSourceUrl,
-      clientIp,
-      clientUserAgent,
-      pixelIdOverride: sourceCampaign?.metaPixelId || undefined,
-      marketingConsent: capiMarketingConsent,
-    }).catch((err) => {
-      d.logger.error('[CAPI] sendLeadEvent error', { error: err?.message || String(err) });
-    });
-
-    // Meta CAPI CompleteRegistration (quiz funnel). Fired server-side only when
-    // the browser sent a registrationEventId (the quiz reveal happened), using
-    // that same id so Meta dedups it against the Pixel CompleteRegistration fired
-    // at the reveal. No-op for non-quiz leads. Guard inside sendCompleteRegistrationEvent.
-    if (registrationEventId) {
-      d.sendCompleteRegistrationEvent(prospect, {
-        eventId: registrationEventId,
-        fbp,
-        fbc,
-        eventSourceUrl,
-        clientIp,
-        clientUserAgent,
-        pixelIdOverride: sourceCampaign?.metaPixelId || undefined,
-        marketingConsent: capiMarketingConsent,
-      }).catch((err) => {
-        d.logger.error('[CAPI] sendCompleteRegistrationEvent error', { error: err?.message || String(err) });
-      });
-    }
-
-    // TikTok Events API dispatch (fire-and-forget; post-commit; guard inside the
-    // sender). Mirrors the Meta CAPI pair: a Lead at submit, plus a
-    // CompleteRegistration when the quiz reveal fired one — each deduped against
-    // the browser ttq pixel via the shared event ids. Per-campaign tiktokPixelId
-    // overrides env TIKTOK_PIXEL_ID.
-    const tiktokCtxBase = {
       ttclid,
       ttp,
-      eventSourceUrl,
       clientIp,
       clientUserAgent,
-      pixelIdOverride: sourceCampaign?.tiktokPixelId || undefined,
-      marketingConsent: capiMarketingConsent,
-    };
-    d.sendTikTokLeadEvent(prospect, { eventId, ...tiktokCtxBase }).catch((err) => {
-      d.logger.error('[TikTok] sendTikTokLeadEvent error', { error: err?.message || String(err) });
     });
-    if (registrationEventId) {
-      d.sendTikTokCompleteRegistrationEvent(prospect, { eventId: registrationEventId, ...tiktokCtxBase }).catch((err) => {
-        d.logger.error('[TikTok] sendTikTokCompleteRegistrationEvent error', { error: err?.message || String(err) });
-      });
-    }
-
-    // Redeem Ops reward-entitlement hook — post-commit, fire-and-forget (a
-    // Redeem Ops failure must never fail or slow lead capture). No-op unless
-    // bootstrap registered the callback (module flag on). Idempotent downstream
-    // via the unique (activationId, prospectId) anchor.
-    // Screening holds ARE reward-eligible (plan D8): the consumer earned the
-    // signup reward by verified signup — screening gates AGENT delivery, not
-    // consumer rewards. Quota/DNC/external holds stay excluded as before.
-    // The hook's ISSUANCE result is kept as an always-resolving promise for the
-    // caller: the controller chains the confirmation email on it so a queued
-    // merged draw email (drawEmailQueued) suppresses the near-duplicate generic
-    // confirmation — an observed outcome, never a prediction, so a skip can
-    // never orphan a signup with zero emails. Still fire-and-forget for the
-    // capture path itself (errors resolve to null, logged as before).
-    let leadCapturedOutcome = null;
-    if (!quarantined || heldReason === 'screening_pending') {
-      try {
-        const hookResult = d.onLeadCaptured?.(prospect);
-        if (hookResult && typeof hookResult.then === 'function') {
-          leadCapturedOutcome = hookResult.catch((err) => {
-            d.logger.error('[RedeemOps] onLeadCaptured error', { error: err?.message || String(err) });
-            return null;
-          });
-        } else if (hookResult != null) {
-          leadCapturedOutcome = Promise.resolve(hookResult);
-        }
-      } catch (err) {
-        d.logger.error('[RedeemOps] onLeadCaptured error', { error: err?.message || String(err) });
-      }
-    }
-
-    // Enrichment: opportunistic post-commit drain of the map-job outbox —
-    // fire-and-forget (the job row is already durable; the sweep re-drains).
-    d.drainMapJobs({ limit: 5 }).catch((err) => {
-      d.logger.warn('[enrichment] post-capture drain failed', { error: err?.message || String(err) });
-    });
-
-    // Pre-load agent + prospect-with-campaign for the caller's fire-and-forget
-    // email side-effects. The campaign's design_config.customerHost drives the
-    // confirmation-email brand, so load the campaign (with design_config) for
-    // EVERY prospect — not only when an agent is assigned.
-    let assignedAgent = null;
-    if (assignedAgentId) {
-      assignedAgent = await m.User.findByPk(assignedAgentId);
-    }
-    const prospectWithCampaign =
-      (await m.Prospect.findByPk(prospect.id, {
-        include: [{ association: 'campaign', attributes: ['id', 'name', 'design_config'] }],
-      })) || prospect;
-
-    // Mint the prospect's ONE canonical referral share link now, on the campaign's
-    // canonical host, so the confirmation email and the SPA share dialog hand out the
-    // identical /share/{slug}. Non-blocking: a failure must never break lead creation —
-    // the email + SPA fall back to the long ?ref= URL. Injected via deps so DI unit tests
-    // can stub it (keeps them DB-free).
-    let shareUrl = null;
-    const shareCampaignId = prospect.campaignId;
-    if (shareCampaignId) {
-      try {
-        const hostChoice = normalizeCustomerHostChoice(
-          prospectWithCampaign?.campaign?.design_config?.customerHost
-        );
-        const origin = customerHostOrigin(hostChoice);
-        const { url } = await d.getOrCreateProspectShareLink({
-          prospectId: prospect.id,
-          campaignId: shareCampaignId,
-          origin,
-        });
-        shareUrl = `${origin}${url}`;
-      } catch (err) {
-        d.logger.warn('Referral share link mint failed (non-blocking)', {
-          prospectId: prospect.id,
-          err: err?.message,
-        });
-      }
-    }
-
-    // AI screening dial trigger (plan §5.3) — LAST on purpose: after the CAPI
-    // dispatches (capture-time events must see the prospect exactly as today)
-    // and the entitlement hook. Fire-and-forget; any failure leaves the held
-    // row for the sweep. DNC-held leads are NOT triggered here — dncGate hands
-    // off and dials on clear (§6).
-    // SCHEDULE, don't dial: the first attempt waits cfg.dialDelaySeconds so the
-    // consumer can read the success page's "an automated call is coming" notice
-    // before their phone rings (§7.1a).
-    if (screeningHeld) {
-      d.scheduleScreeningAttempt(prospect, { campaign: sourceCampaign }).catch((err) =>
-        d.logger.error('[Screening] capture dial trigger error', { error: err?.message || String(err) })
-      );
-    }
 
     return { prospect, assignedAgentId, assignedAgent, prospectWithCampaign, quarantined, shareUrl, leadCapturedOutcome };
   }
